readme update and refactored example app
