+++ at_contacts_flutter/CHANGELOG.md
@@ -1,66 +1,66 @@
 ## 4.0.0
-- Exposed contact services
-- Updated return type of contact services
-- Added documentation for contact service functions
+- **FEAT**: Exposed contact services
+- **CHORE**: Updated return type of contact services
+- **DOCS**: Added documentation for contact service functions
 
 ## 3.1.3
-- Added Package documentation.
-- Added a new example.
-- Added example documentation.
+- **DOCS**: Added Package documentation.
+- **CHORE**: Added a new example.
+- **DOCS**: Added example documentation.
 
 ## 3.1.2
-- Fixed at_client dependency issue
-- Updated at_contact dependency
+- **FIX**: Fixed at_client dependency issue
+- **CHORE**: Updated at_contact dependency
 
 ## 3.1.1
-- Updated package dependencies
+- **CHORE**: Updated package dependencies
 
 ## 3.1.0
 - **FEAT**: Added platform support for macos, windows and linux
 
 ## 3.0.2
-- Updated dependencies and bug fixes
-- Updated documentation
+- **FIX**: Bug fixes
+- **DOCS**: Updated documentation
 
 ## 3.0.1
-- Updated dependencies and bug fixes
+- **CHORE**: Updated dependencies and bug fixes
 
 ## 3.0.0
-- package updated with rsdk changes
+- **CHORE**: package updated with rsdk changes
 
 ## 2.0.3
-- Updated public persona keys
+- **CHORE**: Updated public persona keys
 
 ## 2.0.2
-- Updated repository link and description
-- Updated dependencies
+- **CHORE**: Updated repository link and description
+- **CHORE**: Updated dependencies
 
 ## 2.0.1
-- Updated dependencies
+- **CHORE**: Updated dependencies
 
 ## 2.0.0
-- Null safe version
-- Updated dependencies
+- **CHORE**: Null safe version
+- **CHORE**: Updated dependencies
 
 ## 1.0.2
-- Fix to prevent adding self to contact list
-- Added function to get contact details from cached data
+- **FIX**: Fix to prevent adding self to contact list
+- **FEAT**: Added function to get contact details from cached data
 
 ## 1.0.1
-- The referenced packages are updated
-- Improved avatar placeholder
+- **CHORE**: The referenced packages are updated
+- **CHORE**: Improved avatar placeholder
 
 ## 1.0.0
-- Updated example app
+- **CHORE**: Updated example app
 
 ## 0.0.4
-- The referenced packages are updated
+- **CHORE**: The referenced packages are updated
 
 ## 0.0.3
-- The referenced packages from client sdk updated
+- **CHORE**: The referenced packages from client sdk updated
 
 ## 0.0.2
-- The referenced packages from client sdk updated
+- **CHORE**: The referenced packages from client sdk updated
 
 ## 0.0.1
 - The initial release consists of two screens - contacts and blocked contacts.
