+++ at_contacts_flutter/analysis_options.yaml
@@ -2,9 +2,8 @@ include: package:flutter_lints/flutter.yaml
 
 analyzer:
   exclude: [build/**]
-  # strong-mode:
-  #   implicit-casts: false
 
 linter:
   rules:
     - camel_case_types
+    - unnecessary_statements
