+++ at_contacts_flutter/example/.env
@@ -1,2 +1,1 @@
 ROOT_DOMAIN=root.atsign.org
-NAMESPACE=at_contacts_eg
