+++ at_contacts_flutter/example/pubspec.yaml
@@ -1,4 +1,4 @@
-name: eg
+name: at_contacts_flutter_example
 description: A new Flutter project.
 
 # The following line prevents the package from being accidentally published to
