+++ at_contacts_flutter/lib/at_contacts_flutter.dart
@@ -4,3 +4,4 @@ export 'screens/contacts_screen.dart';
 export 'screens/blocked_screen.dart';
 export 'utils/init_contacts_service.dart';
 export 'desktop_screens/desktop_contacts_screen.dart';
+export 'package:at_common_flutter/at_common_flutter.dart';
